업로드 최종 수정 중: 크롭 완료, 버그 수정중
